Changing how the parser works
As we found out earlier, some of the uploaders upload the "<<<<<< EOF" section like this

```
...
covered="10"/><counter type="CLASS" missed="0" covered="3"/></package><counter type="INSTRUCTION" missed="14" covered="212"/><counter type="LINE" missed="0" covered="32"/><counter type="COMPLEXITY" missed="2" covered="21"/><counter type="METHOD" missed="2" covered="21"/><counter type="CLASS" missed="0" covered="7"/></report><<<<<< EOF
```

at the end of the line, which was not expected, and also not documented on https://docs.codecov.io/v4.4.0/reference#upload (alhtough it's a very lightly defined format there)

The code that deals with that is a little too slow. So we are changing how it works. We are going to lose some flexibility over loading from files. But this is better for the medium-term.