Removing flags section from checks
We are currently having performance problems on notifier. This seems to be made worse when checks was implemented

One of the problems seem to be that checks has to calculate again the same sections that the comment notifier does.

On most section this is not a problem. But it looks like the `flags` section takes up to 30s to calculate. This could be very bad if it is calculated again

Since users don't really have an expectation for this content to be in the checks, we are just removing it for now