Initial RIBS installation
This commit only installs ribs ( https://github.com/codecov/ribs ) in the system, with no code changes to use it

The idea here is just to make sure the Rust-related tooling works well in production

Due to the nature of nested private repos (ribs is inside shared, which is here), the old approach of using https and calling `sed` on the requirements.txt doesn't work anymore, since we can't call sed inside the shared/setup.py. Sp we are changing to an approach where the proper SSH key is already installed at circleci, and it is capable of authenticating every git+ssh requirement inside the `requirements.txt` file. It is even cleaner since it gets local and CI images closer to each other on installation