Changing when the pull.head is updated
After some anecdotal cases happening to myself, I noticed that doing commit-amend and force-push leads to a non-consistent update of the codecov pull request page for the PR involving that commit.

Sometimes I will see the amended commit coverage, sometimes I see the previous commit coverage.

Investigating, I noticed that this line (the one I am changing) is the culprit:

```
head.timestamp < commit.timestamp
```

On itself, the comparison makes sense. If the commit timestamp is higher than the current head than it's a more recent commit on the PR commit tree and it should replace the PR head. The equal case doesn't really need to be covered, since generating two commits at the exact same time is unlikely, even more unlikely on the same PR.

Except for the fact that `timestamp` is the "commit author" date ( https://stackoverflow.com/questions/11856983/why-git-authordate-is-different-from-commitdate ). And amending an existing commit doesn't change its author date. Therefore, every time a commit is amended and force-pushed, it will keep the exact same date as the unamended commit, and we won't set the new commit (commits are immutable) as the PR head. So the user will keep seeing the non-amended commit coverage on the UI (since the UI uses `pull.head` to decide what to show).

So I am making it so same author dates will also change what the PR head is.

While this changes some date-related behavior, I don't think it breaks any flows. Will still ping product for making sure we don't screw up.