Changing how Report is structured
Currently, we have a `Report` with three pieces of structure:

1) Session mapping (which allows to say which flags are part of which coverages)
2) Filename mapping (which allows to say which filenames are the right ones for each given chunk in the array of chunks in chunks.txt)
3) Actual `ReportFile` array, which we use 2 to tell which is the filename of each member in the array

If you have noticed, 2 + 3 have some dependency. If for example the array has 20 elements, but there isn't a filename mapping to element 12, there is a discrepancy. If a filename maps to an element 34 on the same array, that is also a discrepancy.

Truth is, the chunk numbers aren't that relevant, they only serve as a mapping from database to chunks.txt, and once inside a report, they are actually replaced by filenames. so we should use a structure that matches that (specially in rust, which require so many unwrapping steps if you know something will be right anyway).

One could argue that an array of `ReportFile` is still a decent solution, given that the structure is not sparse (which usually calls for usage of HashMaps), and it is nice to split the `ReportFile` themselves from the filenames, given that in some situations the filenames are indeed not relevant.

The problem is that, when a filename is deleted from a Report for some reason (like in carryforward flags or in path resolving), the chunks for that file are not removed, but instead made to 'null'. So we have many situations where a chunk is there but there will indeed be no matching filename. This leads to a situation where the array must contain also `None`s, and it could become a sparse array, nullifying some of the benefits of keeping an array instead of a filename.