Transforming list of Result into Result of list
There is this way in rust ( https://doc.rust-lang.org/rust-by-example/error/iter_result.html#fail-the-entire-operation-with-collect ) where you can turn Vec<Result<A,B>> in a Result<Vec<A>,B> making it so, if one element is an error, the whole result is an errors.

And this is basically what we want in the parsing step.

This will also make future changes easier