Stopping calls to SetPending task
Setpending task is not only a feature that never made it out of beta, but also its implementation (which was ported from legacy) does not seem very proper to its use case: scheduling a task for a setpending might be affected multiple race conditions (and not unlikely ones) where the setpending does not happen soon enough or might even happen after the actual status is posted.

So we are stopping calls to it for now. And will remove the actual task in a future deployment