Using new fields available on the YAML
The goal here is to allow customers to do things like

```
flag_management:
  default_rules:
    statuses:
      - name_prefix: aaa
        type: patch
```

(which is already defined in the YAML for validation)

and enable them to get one status of patch for each flag.

For people who are not using flag_management, we still will send the same flags as before.

Note that this makes it really easier for customers to get 20+ statuses with very little configuration