Fixing pagination header usage on Gitlab
There seem to be times where Gitlab will return a response where the header `"X-Next-Page"` present, but as an empty string. They do that indicating there are no more pages to be returned

We are unsure whether that's common or part of the spec. But regardless we want to make sure we properly stop paginating when we see a false-y value on the `"X-Next-Page"` header