Trying to deal with DeleteOwnerTask timeouts
Here is what we have as indexes and references on the production table:

```
Indexes:
    "owners_pkey" PRIMARY KEY, btree (ownerid)
    "owner_service_ids" UNIQUE, btree (service, service_id)
    "owner_service_username" UNIQUE, btree (service, username)
Foreign-key constraints:
    "owners_bot_fkey" FOREIGN KEY (bot) REFERENCES owners(ownerid) ON DELETE SET NULL
Referenced by:
    TABLE "owners" CONSTRAINT "owners_bot_fkey" FOREIGN KEY (bot) REFERENCES owners(ownerid) ON DELETE SET NULL
    TABLE "sessions" CONSTRAINT "sessions_ownerid_fkey" FOREIGN KEY (ownerid) REFERENCES owners(ownerid) ON DELETE CASCADE
    TABLE "repos" CONSTRAINT "repos_bot_fkey" FOREIGN KEY (bot) REFERENCES owners(ownerid) ON DELETE SET NULL
    TABLE "repos" CONSTRAINT "repos_ownerid_fkey" FOREIGN KEY (ownerid) REFERENCES owners(ownerid) ON DELETE CASCADE
    TABLE "pulls" CONSTRAINT "pulls_author_fkey" FOREIGN KEY (author) REFERENCES owners(ownerid) ON DELETE SET NULL
    TABLE "commits" CONSTRAINT "commits_author_fkey" FOREIGN KEY (author) REFERENCES owners(ownerid) ON DELETE SET NULL
    TABLE "yaml_history" CONSTRAINT "yaml_history_ownerid_fkey" FOREIGN KEY (ownerid) REFERENCES owners(ownerid) ON DELETE CASCADE
    TABLE "yaml_history" CONSTRAINT "yaml_history_author_fkey" FOREIGN KEY (author) REFERENCES owners(ownerid) ON DELETE CASCADE
Triggers:
    owner_cache_state_insert AFTER INSERT ON owners FOR EACH ROW EXECUTE PROCEDURE owner_cache_state_update()
    owner_cache_state_update AFTER UPDATE ON owners FOR EACH ROW WHEN (new.organizations IS DISTINCT FROM old.organizations) EXECUTE PROCEDURE owner_cache_state_update()
    owner_token_clered AFTER UPDATE ON owners FOR EACH ROW WHEN (new.oauth_token IS DISTINCT FROM old.oauth_token AND new.oauth_token IS NULL) EXECUTE PROCEDURE owner_token_clered()
    owner_yaml_updated BEFORE UPDATE ON owners FOR EACH ROW WHEN (((new.yaml -> 'codecov'::text) ->> 'bot'::text) IS DISTINCT FROM ((old.yaml -> 'codecov'::text) ->> 'bot'::text) OR ((new.yaml -> 'codecov'::text) ->> 'branch'::text) IS DI
    owners_before_insert BEFORE INSERT ON owners FOR EACH ROW EXECUTE PROCEDURE owners_before_insert_or_update()
    owners_before_update BEFORE UPDATE ON owners FOR EACH ROW WHEN (new.username IS NOT NULL AND new.username IS DISTINCT FROM old.username) EXECUTE PROCEDURE owners_before_insert_or_update()
```

So the references that need to be cleaned and are not yet are:
- `sessions.ownerid`
- `repos.bot`
- `yaml_history.ownerid`
- `yaml_history.author`

The table yaml_history has currently ~7500 rows, so we are not going to make special query for it and hope it is actually quick on its own (it already has the DELETE CASCADE). For the other references, we are adding queries to delete them.

Also adding retries and a longer time limit for those tasks to run

I believe the longer query is setting bots NULL on the repos.bot query, since it's a big table (currently ~11M rows) and there is no index covering this query. Unfortunately we can either add this index whose purpose will only be User deletion (which is a pretty narrow use case) or have to increase timeouts on this task and add retries. It doesn't even look like a retry will help a lot, since it's a query that takes long regardless of how many positive results it finds:

```
thiago-ro@34:postgres> SELECT * FROM repos WHERE bot=2691665;
...
SELECT 0
Time: 10.472s (10 seconds), executed in: 10.456s (10 seconds)
```

The thing is that, by deleting the user, it will log them out. It is possible the user will get the wrong impression that they were deleted even if this task fails.