Reducing what we consider a discrepancy in Rust totals calculator
Whenever we calculate totals we want to log whenever there is a difference between what rust calculates and what python calculates.

Whenever rust returns `None` as the coverage and python returns `100`, this means that there are no lines covered and python is considering is as a 100% coverage

Python is not wrong. But Rust is also not wrong in this case. So we are making this change so this case won't be logged so we concentrate on actual discrepancies on results.