Using different head
There are a lot of situations where the PR information we have in our system does not exactly matches what Github has.

One example is PR head. On certain occasions, whenever a commit gets pushed to a PR, but doesn't get its coverage uploaded to codecov, our PR head will be the old commit, while the Github PR head will be a different one.

While in a lot of situations the "old PR is good enough", there are a couple of cases, like merging master into their feature branch, makes it so the PR has a more recent base, incompatible with the head with have in our system (think things like the head is from a week ago and more recent master is from today). And this discrepancy might make the user think the wrong things changed.