Setting some situations where coverage is undefined
In our code, we have that the current totals calculations almost everywhere is:

```
coverage = number of hits / number of lines
```

where usually

```
number of lines = number of hits + number of misses + number of partials
```

But when `number of lines = number of hits = 0`, the above calculation is undefined (0 / 0)

Our code currently does `coverage = 100` or `coverage = 0` on those cases, depending on whether the report has files or not.

This behavior is weird, since 100 and 0 are valid coverage results of their own. And this might generate unexpected statuses, like a specific report with a certain flag on 90% coverage will have lower coverage than a report where that same flag is inexistent.

Unfortunately, this change will require some work at the worker and API to properly use this and not break. On the worker, it might be worth it to create specific status messages for the coverage=None result.

But I think in the long term this is the right flow now