Implementing email collection at Github
On Github there is an API endpoint specially for email fetching. Legacy already used this, and it just got missed at the porting

So we are adding it to the `get_authenticated_user` method.

Due to this being `shared`, and we needing a quick turnaround on this issue, and API being currently at shared@v0.4.12, we are branching off that tag to do this change, and will probably just tag the resulting commit as v0.4.12.1 . shared is at v0.4.15 now and we don't want to handle all the other upgrades that need to be done in order to get this change

I am calling this a hotfix for that purpose. But this PR will be merged at the normal shared master branch. And whenever we have more time, API will be properly upgraded to shared v0.4.16, which will have this fix as well