Treating numbers as not jsons during report processing
We have had a couple of situations where the bash uploader, due to its selection of files to upload, will upload files where the content is just a number.

While those files shouldn't be uploaded, we should not crash our system whenever dealing with them

Treating them as jsons break a few of the processors that are expecting dictionaries or arrays.