Starting to use ∅ instead of ø on some of the null cases
As we talked in previous commits, we now have cases where head or base coverage can be null inside of totals.

But in some cases out comments might get confusing because our logic uses ø as the symbol for both null coverage and zero coverage.

While we want to fix that, finding a very different symbol might confuse people.

So for now we are picking a similar symbol (∅), which is the https://en.wikipedia.org/wiki/Null_sign for our nulls. Not all nulls are going to be it yet (ideally the diff coverage being null should be, but it is already the previous ø and it happens a lot already, so we don't want to break that).

Soon someone should consider different symbols when redesigning this.