Applying repack on the report at the time of saving it
Due to the nature of carryforward, we sometimes have a lot of files deleted during the "add this session S to this report", because this leads to sessions being deleted, and some files that only contain those sessions are also deleted for integrity purposes.

Each file deletion leads to a value being replaced with a `None` in the chunks (which is an array). And whenever adding new files, our logic for now is unable of reusing those `None` spaces as places for the files to come.

In some situations, this leads to very big arrays where a minority of the values are actually not None

The `repack` function from shared has the purpose of rearranging the chunks array in a way to make the chunks array more compact

This leads to both better readability of the chunks contents and probably some speedup at parsing the `chunks.txt` file