Making coverage null not make it into the table
Turns out there is at least one case where NULL coverage can make it into the table: empty reports. If a user uploads a report and the end result, after merging everything that already exists with the new results, is an empty report, then the coverage can be null.

We are fixing at https://github.com/codecov/codecov.io/pull/944/files but since we don't have a timeline on that, and also until we know that frontend is ready to deal with this, we will save those nulls as zeroes.