Clearing cache when ignorelines is called
We found that whenever `ignore_lines` is called, the `file._totals` is properly reset. But the `report._files` dict, which contains a cache of the previous result, is not reset.

So we are making so it gets reset. It probably will make the processing slightly longer, but it probably is fine (correctness at this matter is more important)

This needs to be corrected because we want to use those totals in the future to be able to reuse pre-calculations